Add IdentityDbContext migration with initial data
This commit introduces a new migration for the `IdentityDbContext` in the `RentEase.Identity` namespace. It creates several tables related to ASP.NET Core Identity, including `AspNetRoles`, `AspNetUsers`, `AspNetRoleClaims`, `AspNetUserClaims`, `AspNetUserLogins`, `AspNetUserRoles`, and `AspNetUserTokens`. Initial data for roles and users is also inserted, and the model snapshot file is updated to reflect the current database schema.